start to extract all Entry Tests
