+++ budget-binder-server/src/test/kotlin/de/hsfl/budgetBinder/server/EntryTest.kt
@@ -124,162 +124,6 @@ class EntryTest {
 
 
     @Test
-    fun testCreateEntry() = customTestApplicationWithLogin { client ->
-        client.get("/entries").let { response ->
-            assertEquals(HttpStatusCode.Unauthorized, response.status)
-            val responseBody: APIResponse<Entry> = response.body()
-            val shouldResponse: APIResponse<Entry> = wrapFailure("Your accessToken is absent or does not match.", 401)
-            assertEquals(shouldResponse, responseBody)
-        }
-
-        client.sendAuthenticatedRequest(HttpMethod.Post, "/entries") { response ->
-            assertEquals(HttpStatusCode.OK, response.status)
-            val responseBody: APIResponse<Entry> = response.body()
-            val shouldResponse: APIResponse<Entry> = wrapFailure("The object you provided it not in the right format.")
-            assertEquals(shouldResponse, responseBody)
-        }
-
-        client.sendAuthenticatedRequestWithBody(
-            HttpMethod.Post, "/entries",
-            Entry.In("Bafög", 700f, true, null)
-        ) { response ->
-            assertEquals(HttpStatusCode.OK, response.status)
-            val responseBody: APIResponse<Entry> = response.body()
-
-            val id = transaction {
-                EntryEntity.all().last().let {
-                    assertEquals("Bafög", it.name)
-                    assertEquals(700f, it.amount)
-                    assert(it.repeat)
-                    assertNull(it.category)
-                    it.id.value
-                }
-            }
-            val shouldResponse = wrapSuccess(Entry(id, "Bafög", 700f, true, null))
-            assertEquals(shouldResponse, responseBody)
-        }
-    }
-
-
-    @Test
-    fun testGetEntries() = customTestApplicationWithLogin { client ->
-        client.get("/entries").let { response ->
-            assertEquals(HttpStatusCode.Unauthorized, response.status)
-            val responseBody: APIResponse<List<Entry>> = response.body()
-            val shouldResponse: APIResponse<List<Entry>> = wrapFailure("Your accessToken is absent or does not match.", 401)
-            assertEquals(shouldResponse, responseBody)
-        }
-
-        val id = transaction { EntryEntity.all().first().id.value }
-
-        val entryList = listOf(
-            Entry(id, "Monthly Job Pay", 3000f, true, null),
-            Entry(id + 1, "Monthly Job Pay", 3500f, true, null),
-            Entry(id + 2, "Phone", -50f, true, null),
-            Entry(id + 3, "Internet", -50f, true, null),
-            Entry(id + 4, "Bike", -1500f, false, null),
-            Entry(id + 5, "Ikea", -200f, false, null),
-            Entry(id + 6, "new Phone", -50f, true, null),
-        )
-
-        client.sendAuthenticatedRequest(HttpMethod.Get, "/entries") { response ->
-            assertEquals(HttpStatusCode.OK, response.status)
-            val responseBody: APIResponse<List<Entry>> = response.body()
-            assert(responseBody.success)
-            assertEquals(7, responseBody.data!!.size)
-            val shouldResponse = wrapSuccess(entryList)
-            assertEquals(shouldResponse, responseBody)
-        }
-
-        client.sendAuthenticatedRequest(HttpMethod.Get, "/entries?current=true") { response ->
-            assertEquals(HttpStatusCode.OK, response.status)
-            val responseBody: APIResponse<List<Entry>> = response.body()
-            assert(responseBody.success)
-            assertEquals(4, responseBody.data!!.size)
-
-            val currentList = listOf(entryList[1], entryList[3], entryList[5], entryList[6])
-            val shouldResponse = wrapSuccess(currentList)
-            assertEquals(shouldResponse, responseBody)
-        }
-
-        client.sendAuthenticatedRequest(HttpMethod.Get, "/entries?period=50-8346") { response ->
-            assertEquals(HttpStatusCode.OK, response.status)
-            val responseBody: APIResponse<List<Entry>> = response.body()
-            val shouldResponse: APIResponse<List<Entry>> = wrapFailure("period has not the right pattern")
-            assertEquals(shouldResponse, responseBody)
-        }
-
-        val now = LocalDateTime.now()
-
-        client.sendAuthenticatedRequest(HttpMethod.Get, "/entries?period=${formatToPeriod(now)}") { response ->
-            assertEquals(HttpStatusCode.OK, response.status)
-            val responseBody: APIResponse<List<Entry>> = response.body()
-            assert(responseBody.success)
-            assertEquals(4, responseBody.data!!.size)
-
-            val currentList = listOf(entryList[1], entryList[3], entryList[5], entryList[6])
-            val shouldResponse = wrapSuccess(currentList)
-            assertEquals(shouldResponse, responseBody)
-        }
-
-        client.sendAuthenticatedRequest(
-            HttpMethod.Get,
-            "/entries?period=${formatToPeriod(now.minusMonths(1))}"
-        ) { response ->
-            assertEquals(HttpStatusCode.OK, response.status)
-            val responseBody: APIResponse<List<Entry>> = response.body()
-            assert(responseBody.success)
-            assertEquals(3, responseBody.data!!.size)
-
-            val currentList = listOf(entryList[1], entryList[3], entryList[4])
-            val shouldResponse = wrapSuccess(currentList)
-            assertEquals(shouldResponse, responseBody)
-        }
-
-        client.sendAuthenticatedRequest(
-            HttpMethod.Get,
-            "/entries?period=${formatToPeriod(now.minusMonths(2))}"
-        ) { response ->
-            assertEquals(HttpStatusCode.OK, response.status)
-            val responseBody: APIResponse<List<Entry>> = response.body()
-            assert(responseBody.success)
-            assertEquals(3, responseBody.data!!.size)
-
-            val currentList = listOf(entryList[1], entryList[2], entryList[3])
-            val shouldResponse = wrapSuccess(currentList)
-            assertEquals(shouldResponse, responseBody)
-        }
-
-        client.sendAuthenticatedRequest(
-            HttpMethod.Get,
-            "/entries?period=${formatToPeriod(now.minusMonths(3))}"
-        ) { response ->
-            assertEquals(HttpStatusCode.OK, response.status)
-            val responseBody: APIResponse<List<Entry>> = response.body()
-            assert(responseBody.success)
-            assertEquals(2, responseBody.data!!.size)
-
-            val currentList = listOf(entryList[0], entryList[2])
-            val shouldResponse = wrapSuccess(currentList)
-            assertEquals(shouldResponse, responseBody)
-        }
-
-        client.sendAuthenticatedRequest(
-            HttpMethod.Get,
-            "/entries?period=${formatToPeriod(now.minusMonths(4))}"
-        ) { response ->
-            assertEquals(HttpStatusCode.OK, response.status)
-            val responseBody: APIResponse<List<Entry>> = response.body()
-            assert(responseBody.success)
-            assertEquals(0, responseBody.data!!.size)
-
-            val shouldResponse: APIResponse<List<Entry>> = wrapSuccess(emptyList())
-            assertEquals(shouldResponse, responseBody)
-        }
-    }
-
-
-    @Test
     fun testGetEntryById() = customTestApplicationWithLogin { client ->
         client.get("/entries/1").let { response ->
             assertEquals(HttpStatusCode.Unauthorized, response.status)
