+++ budget-binder-server/src/test/kotlin/de/hsfl/budgetBinder/server/entries/GetEntryByIDTests.kt
@@ -1,0 +1,41 @@
+package de.hsfl.budgetBinder.server.entries
+
+import de.hsfl.budgetBinder.server.models.EntryEntity
+import de.hsfl.budgetBinder.server.models.UserEntity
+import de.hsfl.budgetBinder.server.utils.customTestApplication
+import de.hsfl.budgetBinder.server.utils.registerUser
+import org.jetbrains.exposed.sql.transactions.transaction
+import java.time.LocalDateTime
+import kotlin.test.*
+
+class GetEntryByIDTests {
+
+    @BeforeTest
+    fun before() = customTestApplication { client ->
+        client.registerUser()
+
+        val userEntity = transaction { UserEntity.all().first() }
+        val now = LocalDateTime.now()
+
+        transaction {
+            EntryEntity.new {
+                name = "Monthly Pay"
+                amount = 3000f
+                repeat = true
+                created = now.minusMonths(3)
+                ended = now.minusMonths(2)
+                child = null
+
+                user = userEntity
+                category = null
+            }
+        }
+    }
+
+    @AfterTest
+    fun after() = transaction {
+        UserEntity.all().forEach { it.delete() }
+    }
+
+
+}
